displayEmployee function added
the display employee function is added and seems to be working. the entire code is to be able to compile and run. However when option 1 is selected, and  employee1002 is entered. "Invalid option, try again" is displayed. It seem to have to do with something in the findEmployee function. But I'm not entirely sure.